move tracer to its own file
