+++ cmd/ef_tests/state/runner/levm_runner.rs
@@ -12,10 +12,7 @@ use ethrex_common::{
     H256, U256,
 };
 use ethrex_levm::{
-    db::gen_db::GeneralizedDatabase,
-    errors::{ExecutionReport, TxValidationError, VMError},
-    vm::{LevmCallTracer, VM},
-    EVMConfig, Environment,
+    calltracer::LevmCallTracer, db::gen_db::GeneralizedDatabase, errors::{ExecutionReport, TxValidationError, VMError}, vm::VM, EVMConfig, Environment
 };
 use ethrex_rlp::encode::RLPEncode;
 use ethrex_vm::backends;
@@ -191,7 +188,7 @@ pub fn prepare_vm_for_tx<'a>(
         },
         db,
         &tx,
-        LevmCallTracer::disabled(),
+        LevmCallTracer::new(false, true),
     ))
 }
