+++ crates/vm/levm/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod call_frame;
+pub mod calltracer;
 pub mod constants;
 pub mod db;
 pub mod environment;
